chore(suite-native): move pin request to device-authorization
