+++ suite-native/app/src/navigation/AppTabNavigator.tsx
@@ -3,7 +3,7 @@ import { useSelector } from 'react-redux';
 import { BottomTabBarProps, createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 
 import { EventType, analytics } from '@suite-native/analytics';
-import { useHandleDeviceRequestsPassphrase } from '@suite-native/device-authorization';
+import { useHandleDeviceAuthorization } from '@suite-native/device-authorization';
 import { AccountsStackNavigator } from '@suite-native/module-accounts-management';
 import { HomeStackNavigator } from '@suite-native/module-home';
 import { SettingsScreen } from '@suite-native/module-settings';
@@ -39,7 +39,7 @@ const getTradingAnalyticsType = (
 };
 
 export const AppTabNavigator = () => {
-    useHandleDeviceRequestsPassphrase();
+    useHandleDeviceAuthorization();
 
     const isTradingEnabled = useSelector(selectIsTradingEnabled);
     const isTradingBuyEnabled = useSelector(selectIsTradingBuyEnabled);
