+++ suite-native/device/package.json
@@ -32,7 +32,6 @@
         "@suite-native/atoms": "workspace:*",
         "@suite-native/bluetooth": "workspace:*",
         "@suite-native/config": "workspace:*",
-        "@suite-native/device-authorization": "workspace:*",
         "@suite-native/device-mutex": "workspace:*",
         "@suite-native/device-onboarding": "workspace:*",
         "@suite-native/feature-flags": "workspace:*",
