+++ suite-native/module-onboarding/src/screens/BiometricsScreen.tsx
@@ -48,8 +48,6 @@ export const BiometricsScreen = ({
     const exitOnboardingFlow = () => {
         dispatch(setIsOnboardingFinished());
 
-        // TODO: COSMETIC IMPROVEMENT: redirect to home only if there is no device connected. In case of device connected,
-        // the redirect is handled in useHandleDeviceConnection hook. in reaction to the `setIsOnboardingFinished` call.
         navigation.dispatch(
             CommonActions.reset({
                 index: 0,
