+++ suite-native/navigation/src/hooks/useLastRouteName.ts
@@ -1,0 +1,3 @@
+import { useNavigationState } from '@react-navigation/native';
+
+export const useLastRouteName = () => useNavigationState(state => state.routes.at(-1)?.name);
